% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peptide_FDR.R
\name{compute_FDR}
\alias{compute_FDR}
\title{Calculate peptide ID False Discover Rate}
\usage{
compute_FDR(main.pescal.dir, decoy.pescal.dir, peptides)
}
\arguments{
\item{peptides}{}
}
\description{
Calculate peptide ID False Discover Rate
}
