% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_testing.R
\name{labelled_protools2_limma}
\alias{labelled_protools2_limma}
\title{Perform protools2 limma and label output}
\usage{
labelled_protools2_limma(
  df.design,
  df.comparisons,
  df.normalized.areas,
  combipept_ptm
)
}
\arguments{
\item{combipept_ptm}{}
}
\description{
Perform protools2 limma and label output
}
