% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_testing.R
\name{plot_all_volcano}
\alias{plot_all_volcano}
\title{Title}
\usage{
plot_all_volcano(comparisons)
}
\arguments{
\item{comparisons}{List of limma comparisons. Generated from "labelled_protools2_limma".}
}
\description{
Title
}
