% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peptide_FDR.R
\name{plot_mscr_ranks}
\alias{plot_mscr_ranks}
\title{Plot mascot score ranks}
\usage{
plot_mscr_ranks(main.pescal.dir, decoy.pescal.dir)
}
\arguments{
\item{decoy.pescal.dir}{}
}
\description{
Plot mascot score ranks
}
