% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptm_reap.R
\name{ptm_norm}
\alias{ptm_norm}
\title{Title}
\usage{
ptm_norm(output_areas_df)
}
\arguments{
\item{output_areas_df}{Dataframe. Output areas data from pescal output.}
}
\value{
List of dataframes with normalised data. Note, recent focus
has shifted to df.norm.log2.centered.na.impute.v2.
}
\description{
Title
}
