% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptm_reap.R
\name{ptm_reap}
\alias{ptm_reap}
\title{PTM reap}
\usage{
ptm_reap(ptm_pattern, pescal.path, normalise = TRUE, delta_score_cutoff = 10)
}
\arguments{
\item{ptm_pattern}{String. Ptm type to retrieve from combipept data. Must be
exact match.}

\item{pescal.path}{Path to pescal output file.}

\item{normalise}{Logical. Whether to normalise or return raw values.}

\item{delta_score_cutoff}{Numeric. Max_delta_score cut off value for ptm peptides.}
}
\description{
Wrapper function for reading combipept data, output areas, filtering modified
peptides and return normalised results (if selected)
}
