% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathway_enrichment.R
\name{ptmprot_pathway_enrichment}
\alias{ptmprot_pathway_enrichment}
\title{Pathway enrichment of modified peptides}
\usage{
ptmprot_pathway_enrichment(comparisons, combipept, databases, graph.heading)
}
\arguments{
\item{comparisons}{List of comparison dataframes from protools2 limma comparison.}

\item{combipept}{Combipept data from pescal output. Requires whole sheet.}

\item{databases}{Databases to use in pathway enrichment search.}

\item{graph.heading}{String. Title for graphs.}
}
\value{
List of pathway enrichment data from protools2::pathway_enrichment function.
}
\description{
Pathway enrichment of modified peptides. Note, this approach uses protein id
taht modified peptide is assignmed to. It is therefore a proxy of the type of
proteins being modified, and not necessarily representative of the functional
consequences of the modification.
}
