% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptm_reap.R
\name{read_combipept}
\alias{read_combipept}
\title{Read combipept data}
\usage{
read_combipept(pescal.path)
}
\arguments{
\item{pescal.path}{Path with location of of pescal output file.}
}
\value{
Dataframe with combipept data with "acc_no", "protein", "peptide",
"pep_mod", "...25", "max_scr", "mean_scr", "max_delta_score", "xmod_pos" and
"db_id" columns.
}
\description{
Function to read in combipept sheet from pescal output.
}
