% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptm_reap.R
\name{read_output_areas}
\alias{read_output_areas}
\title{Read output areas data}
\usage{
read_output_areas(pescal.path)
}
\arguments{
\item{pescal.path}{Path with location of of pescal output file.}
}
\value{
Dataframe with output areas sheet from pescal output file.
}
\description{
Read output areas data
}
